Rename matcher.js to create-matcher.js
