+++ lib/create-matcher.js
@@ -29,10 +29,27 @@ var matcher = {
     }
 };
 
+/**
+ * Returns `true` when `object` is a matcher
+ *
+ * @private
+ * @param {*} object A value to examine
+ * @returns {boolean} Returns `true` when `object` is a matcher
+ */
 function isMatcher(object) {
     return isPrototypeOf(matcher, object);
 }
 
+/**
+ * Ensures that value is of type
+ *
+ * @private
+ * @param {*} value A value to examine
+ * @param {string} type A basic JavaScript type to compare to, e.g. "object", "string"
+ * @param {string} name A string to use for the error message
+ * @throws {TypeError} If value is not of the expected type
+ * @returns {undefined}
+ */
 function assertType(value, type, name) {
     var actual = typeOf(value);
     if (actual !== type) {
@@ -47,6 +64,16 @@ function assertType(value, type, name) {
     }
 }
 
+/**
+ * Throws a TypeError when expected method doesn't exist
+ *
+ * @private
+ * @param {*} value A value to examine
+ * @param {string} method The name of the method to look for
+ * @param {name} name A name to use for the error message
+ * @param {string} methodPath The name of the method to use for error messages
+ * @throws {TypeError} When the method doesn't exist
+ */
 function assertMethodExists(value, method, name, methodPath) {
     if (value[method] === null || value[method] === undefined) {
         throw new TypeError(
@@ -55,16 +82,37 @@ function assertMethodExists(value, method, name, methodPath) {
     }
 }
 
+/**
+ * Throws a TypeError when `value` is not a matcher
+ *
+ * @private
+ * @param {*} value The value to examine
+ */
 function assertMatcher(value) {
     if (!isMatcher(value)) {
         throw new TypeError("Matcher expected");
     }
 }
 
+/**
+ * Returns `true` for iterables
+ *
+ * @private
+ * @param {*} value A value to examine
+ * @returns {boolean} Returns `true` when `value` looks like an iterable
+ */
 function isIterable(value) {
     return Boolean(value) && typeOf(value.forEach) === "function";
 }
 
+/**
+ * Matches `actual` with `expectation`
+ *
+ * @private
+ * @param {*} actual A value to examine
+ * @param {object} expectation An object with properties to match on
+ * @returns {boolean} Returns true when `actual` matches all properties in `expectation`
+ */
 function matchObject(actual, expectation) {
     if (actual === null || actual === undefined) {
         return false;
@@ -139,6 +187,14 @@ var TYPE_MAP = {
     }
 };
 
+/**
+ * Creates a matcher object for the passed expectation
+ *
+ * @alias module:samsam.createMatcher
+ * @param {*} expectation An expecttation
+ * @param {string} message A message for the expectation
+ * @returns {object} A matcher object
+ */
 function match(expectation, message) {
     var m = Object.create(matcher);
     var type = typeOf(expectation);
@@ -260,6 +316,14 @@ match.instanceOf = function(type) {
     }, "instanceOf(" + (functionName(type) || objectToString(type)) + ")");
 };
 
+/**
+ * Creates a property matcher
+ *
+ * @private
+ * @param {Function} propertyTest A function to test the property against a value
+ * @param {string} messagePrefix A prefix to use for messages generated by the matcher
+ * @returns {object} A matcher
+ */
 function createPropertyMatcher(propertyTest, messagePrefix) {
     return function(property, value) {
         assertType(property, "string", "property");
