+++ lib/create-matcher.test.js
@@ -2,7 +2,7 @@
 
 var assert = require("@sinonjs/referee").assert;
 var refute = require("@sinonjs/referee").refute;
-var createMatcher = require("./matcher");
+var createMatcher = require("./create-matcher");
 
 function propertyMatcherTests(matcher, additionalTests) {
     return function() {
