+++ lib/match.js
@@ -9,7 +9,7 @@ var getClass = require("./get-class");
 var isDate = require("./is-date");
 var isSet = require("./is-set");
 var isSubset = require("./is-subset");
-var createMatcher = require("./matcher");
+var createMatcher = require("./create-matcher");
 
 function arrayContains(array, subset, compare) {
     if (subset.length === 0) {
