+++ lib/samsam.js
@@ -7,7 +7,7 @@ var isNegZero = require("./is-neg-zero");
 var isSet = require("./is-set");
 var match = require("./match");
 var deepEqualCyclic = require("./deep-equal").use(match);
-var createMatcher = require("./matcher");
+var createMatcher = require("./create-matcher");
 
 module.exports = {
     createMatcher: createMatcher,
